+++ src/app/components/forms/terra-forms.example.ts
@@ -1,0 +1,30 @@
+import { Component } from '@angular/core';
+import { TerraSelectBoxValueInterface } from './select-box/data/terra-select-box.interface';
+
+@Component({
+    selector: 'terra-forms-example',
+    template: require('./terra-forms.example.html')
+})
+
+export class TerraFormsExampleComponent
+{
+    protected readonly listBoxValues:Array<TerraSelectBoxValueInterface> = [
+        {
+            caption: '',
+            value:   null
+        },
+        {
+            caption: '1',
+            value:   1
+        },
+        {
+            caption: '2',
+            value:   2
+        }
+    ];
+
+    protected submit():void
+    {
+        // implement submit routine here
+    }
+}